requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
APScheduler==3.11.2
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
beautifulsoup4==4.14.3
cbor2==5.8.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
diff-match-patch==20241021
Django==6.0
django-apscheduler==0.7.0
django-import-export==4.3.14
django-log-viewer==1.1.8
django-stubs==5.2.8
django-stubs-ext==5.2.8
django-unfold==0.75.0
frozenlist==1.8.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
Levenshtein==0.27.3
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
pyTelegramBotAPI==4.29.1
python-decouple==3.8
python-Levenshtein==0.27.3
RapidFuzz==3.14.3
redis==7.1.0
requests==2.32.5
service-identity==24.2.0
soupsieve==2.8
sqlparse==0.5.4
tablib==3.9.0
telebot==0.0.5
thefuzz==0.22.1
Twisted==25.5.0
txaio==25.12.2
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.2
whitenoise==6.11.0
yarl==1.22.0
zope.interface==8.1.1
